Properly rollback propagated states when reprocessing a block
Summary:
During analysis of a block, we may want to "reprocess" the block. This
happens when some transformation is done that makes it necessary to
restart the analysis from the beginning of the block. Generally, all
side-effects that have occurred as part of the analysis must be
undone.

However, we were missing one important part. The analysis may have
propagated state to any number of target blocks before
reprocessing. This is bad because the analysis when we reprocess is no
longer *correct* until we finish the entire block. We were making no
effort to rollback these propagated states in this case.

Fix this by recording the target block's original in-state when
propagating state to it. If we need to reprocess the block we can use
these recorded states to reset those target blocks' states back. This
ensures that the erroneous reprocessed state does not leak out to
other blocks.

Very often the way this would materialize is by failing to reach a
true fixed point. Analysis would finish, but if we ran it again, it
would come up with different results. This breaks invariants so we
cannot allow it.

Reviewed By: mdko

Differential Revision: D56973943

fbshipit-source-id: a5545ea3df09054a3a310ddaf861cfe7724086e1