Track usedParams state on a per-block basis rather than in CollectedInfo
Summary:
Storing usedParams in CollectedInfo causes problems because it's
global to the entire function. If we analyze a block and mark a
parameter as used, then reprocess or rewind that block and analyze it
again, the new analysis may not mark that param as being used, but the
bit stays set! We might think we reached a fixed point, but if we run
the analysis again, we'll get a different result! This breaks
invariants. To fix this, do not track the used param set across the
entire function. Instead track it on a per-block basis. If we
reprocess or rewind the block, the used param set gets cleared. We
only "commit" this per-block used param set to the total end result
once the block is completely processed.

Reviewed By: voorka

Differential Revision: D56972744

fbshipit-source-id: 2fe7d4ae905021645823042f0c00f50e7ce50a16