Add extended reasons for class subtyping
Summary:
Adds extended reasons through class subtyping.

For the case where we have two equal class names this follows same pattern as function subtyping for covariant, contravariant and invariant type parameter projections.

For the case of a subclass, we record the additional flow from the subclass to its common superclass.

Reviewed By: madgen

Differential Revision: D56704136

fbshipit-source-id: 0a2a90a81b44e7b219e649b077d85660ca2c3ddb