Disable existential type stripping
Summary:
Existential type stripping was causing huge regression in files that typecheck against a lot of disjuncts.

This happens because dependent type stripping utility turned out to be using regular expressions which is expensive in a hot code path.

To fix it properly and re-enable stripping, we should either make the computation lazy or even better do it right before the error is presented.

N.b. just because we generate an error doesn't mean that it'll be shown to a user because we'll speculatively try disjuncts and generate errors on failure which might get discarded when a successful disjunct is found.

Reviewed By: enetsee

Differential Revision: D57156170

fbshipit-source-id: 013dc18e721634237d2b245603d20c76c88f80f6