Fix flaky ParallelConcurrencyControllerTest
Summary:
Fix flaky ParallelConcurrencyControllerTest.

**Issue 1:**
The requestCount() is sometimes 1, but it's expected to be 0.

The exact cause of this would be complex to find, but adding simple sleep makes those tests stable.

**Issue 2:**
The LimitedTasks test created two request for `blockingAP`, but it was missing to use `staringBlockingAP`. This was causing `folly::Baton` to throw assertion exceptions.

The fix it to use  `staringBlockingAP` for the third request.

Reviewed By: sazonovkirill

Differential Revision: D57170350

fbshipit-source-id: 6059b085b68e8de0a99707d5a24c184ed060d5d9