Separate out parsing options
Summary:
Move parsing options out of the global options into their own record. This was all of the options prefixed with `po_` (except for `po_disallow_toplevel_requires` which doesn't seem to be used in the parser), as well as`tco_union_intersection_type_hints`, `tco_is_systemlib`, `tco_no_parser_readonly_check`, and `tco_everything_sdt` which are all used in the parser.

Since they are in their own record, drop the `po_` prefix.

Reviewed By: vassilmladenov

Differential Revision: D57119455

fbshipit-source-id: 13785bcf3dd4abe715d358317bbd70675e861eb5