Deshim VirtualExecutor in folly
Summary:
The following rules were deshimmed:
```
//folly:virtual_executor -> //folly/executors:virtual_executor
```

The following headers were deshimmed:
```
folly/VirtualExecutor.h -> folly/executors/VirtualExecutor.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

Reviewed By: Orvid

Differential Revision: D57217665

fbshipit-source-id: 46888473f82ce08f284864778d13d5fdf426b8d2