Refactor Typing.Expr.condition to use type switch constraint more
Summary:
Previously, type switch would only be used if an `is` expression appeared directly as the condition of an if statement but not as part of a compound logical condition which was handled by `Expr.condition`.

`Expr.condition` allowed for updating the environment for each branch but it does so separately. The new type switch constraint involves creating a new environment for both sides in coordination.

This diff refactors `condition` into `condition_dual` which allows for this coordination.

`condition_dual` directly handles expressions that need this kind of coordination as well as logical operators that may contain such expressions. Other expressions are handled by `condition_single` which contains the remainder of the original `condition` handling.

The `if` statement uses `condition_dual` directly. Other callers continue to call `condition` which is now a wrapper around `condition`.

Next, look into:
* Possibly use type switch for some of the cases in condition_single
* Possibly refactor users of new `condition` wrapper to use `condition_dual` directly if advantageous.

Differential Revision: D57053427

fbshipit-source-id: fa88b8e380c790bb915b6aabebfbc487d062011e