Don't crash watchman when detected Eden root is not connected (pt. 3)
Summary: Previous attempt, D56780919, did not fix crash when reading (`.../.eden/socket`). Why? The exception was getting propagated correctly, caused `makeThriftChannel` to propagate that exception eventuanly causing watchman to terminate. Here we catch the `RootNotConnectedError` and return an empty string for the socket address. This results in `makeThriftChannel` failing eventually resulting in `subscriberThread` to exit - which appears to be OK.

Reviewed By: MichaelCuevas

Differential Revision: D57190983

fbshipit-source-id: 7be32db629e2d171b3edf1e04ea627d836f53451