Make typeconst check exhaustive
Summary:
The well-formedness checks for type consts happened as a consequence of localizing the types for subtyping against bounds, and then ad-hoc for concrete type consts where the type is discarded. This change simply makes the match exhaustive, extending the check to bounds by consequence.

Open question if well-formedness should handle unbound type consts directly. As of now, it does _not_ handle unbound names in an Happly.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D57275934

fbshipit-source-id: 9b44cc547e7c8f46847d3d5f301f9173ebd334a2