Deshim bser in watchman
Summary:
The following rules were deshimmed:
```
//folly/experimental/bser:bser -> //folly/json/bser:bser
```

The following headers were deshimmed:
```
folly/experimental/bser/Bser.h -> folly/json/bser/Bser.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

Reviewed By: MichaelCuevas

Differential Revision: D57280129

fbshipit-source-id: 63790675d55327f61d1bcd641e1453d9fa121c2c