Do not add dynamic back on the lower side of the true side if IsNull
Summary:
There's a special handling of like types in the type switch constraint to avoid creating intersections with dynamic. Instead of creating `(dynamic & P) | (T & P)`, we create `~(T & P)`.

While the latter is a less precise type, it improves things in certain cases b/c the typechecker has some logic that doesn't play well with the intersections with dynamic.

However, it has the side-effect of producing `~null` in the true side of `is null`. This type can be problematic and it's actually try that the true side can be precisely `null`.

This diff make this targetted change for now, but we will revisit the decision to not generate `(dynamic & P)`.

Reviewed By: vassilmladenov

Differential Revision: D57333193

fbshipit-source-id: ff06fe69e2b2f2a1f03467398f748586331ba0b3