Error on non-void __clone return
Summary:
it is an occasional misunderstanding that you should return the new object (or anything) from `__clone`. Any returned value is lost as the `__clone` method is meant to be sid-effect-ful.

Possible feedback point: Should we allow `noreturn/nothing` ? We won't recognize `noreturn`-ness at the `clone` operator (callsite), but maybe it's reasonable for a user to express this in their implementation?

Reviewed By: enetsee, madgen

Differential Revision: D57277050

fbshipit-source-id: 6a27ff3e27747ecab674764d3a14d04af64e3d04