Support C++ pointer iteration of vanilla dicts with tombstones
Summary:
Use VanillaDictPointer iterator implementation for iterating const vanilla
dicts with tombstones.

This might look like it is worse, as we have to perform an extra check when
iterating dicts without tombstones, but in practice we iterate almost all such
dicts from jit.

However, the improvement for dicts with tombstones is negligible as well.

The main benefit this brings us is the possibility of determining iterator type
without inspecting the base (beyond the base type), which will allow us to to
kill the nextHelperIdx altogether and improve JIT implementation of IterNext
that will no longer have to match whatever IterInit decided.

Differential Revision: D57359244

fbshipit-source-id: 37f98d04d3a036d430aa5251e5f9b18f804dc3c3