Initial implementation for `MutableSet`
Summary:
This is an initial implementation for the Thrift `MutableSet`. It implements `MutableSet` from the [collections abstract base classes](https://docs.python.org/3.10/library/collections.abc.html#collections-abstract-base-classes). However, not all abstract methods are implemented; these can be completed later.
While the implementation could be optimized, especially for integral types, it is not within the scope of this diff.

Reviewed By: aristidisp

Differential Revision: D57265231

fbshipit-source-id: bcc49f9c5b2f493f12e307539969accd583c383e