Make __fromShape codegen cheaper to typecheck
Summary:
In Hack Thrift codegen, `__fromShape` codegen generates code of the following form:

```
Shapes::idx($shape, 'key') === null ? null : (Foo::__fromShape($shape['key']))
```

It'll generate one of these per constructor parameter which can get quite big in case of compiling large unions.

Each of these refine the type of `$shape` in branches of the ternary expression. In one branch, it'll definitely have the key `'key'` and in the other it'll be optional. On the exist these two types need to be consolidated to obtain what we started with. This can get very expensive.

This diff changes the code generator for this pattern slightly to avoid the definement on `$shape` and consequently the expensive type consolidation.

The new code generated looks as follows:
```
Shapes::idx($shape, 'key') |> $$ === null ? null : (Foo::__fromShape($$))
```

Reviewed By: rmakheja

Differential Revision: D56428792

fbshipit-source-id: 937f8a954df4cf9a4065b4f7a5729d813a6ae07c