Deshim VirtualExecutor in thrift
Summary:
The following rules were deshimmed:
```
//folly:virtual_executor -> //folly/executors:virtual_executor
```

The following headers were deshimmed:
```
folly/VirtualExecutor.h -> folly/executors/VirtualExecutor.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

Reviewed By: vitaut

Differential Revision: D57217657

fbshipit-source-id: f02efff8358461fb61b469b4160059ef0e65447e