Optimise Exit environment handling
Summary:
On return, we move continuations to the `Exit` environment and various control-flow operations require `Exit` continuations to be consolidated.

I believe this is unnecessary most of the time except:
* in try-catch blocks because `finally` will need the environment at the point of `return`
* in lambda bodies because we check if type variables escape their scope

This diff optimises environment merging for the common case where we are outside a try-catch block or a lambda body.

It does so by simply dropping the `Next` continuation instead of stashing it in the `Exit` continuation and consequently doing away with all the environment unioning that happens due to control flow.

Reviewed By: CatherineGasnier

Differential Revision: D57362682

fbshipit-source-id: d3a1203f4e3168f7478dc912e9a178f6e720dd06