reduce unsoundness in comparison binop typechecking
Summary: Checking of comparisons doesn't do anything, when one of the types is a type variable that we don't know anything about yet. This is unsound, since the variable might not turn out to be the right type. Here we insist that each operand must be a subtype of int, string, datetime, or datetimeimmutable no matter what. This still leaves open the possibility of comparing int and string, but not ?int and int.

Reviewed By: vassilmladenov

Differential Revision: D57379513

fbshipit-source-id: 6e811897204e31818cc137075c1e0b8becd2d258