Use array instead of iter slot to perform IterNext specialization checks
Summary:
Check array type and load necessary info from the array to perform IterNext
specialization checks.

Checks that are based on the type will get optimized away whenever CheckIter
previously got optimized away.

CheckArrayKeys won't yet get optimized away, but we do not seem to be
performance sensitive to it. This will be addressed nevertheless by adding
support for JIT iteration over tombstones in the future diffs.

Bespoke tombstone presence checks get optimized away for StructDicts, which
are the only bespoke arrays in use.

This allows removal of iterator type from the Iter slot, which eliminates a
store and reduces Iter slot to 16 bytes.

It also makes it possible for the next diff to specialize IterNext even without
a matching IterInit.

Differential Revision: D57254777

fbshipit-source-id: 776f4aa69bbdf4fe670b917dce88e759c98b9d89