Improve handling of keysets in C++ iteration helpers
Summary:
Use optimized keyset iteration helpers in C++ instead of using the generic
implementation.

Use pointers to VanillaKeysetElm if the base is const.

While this might improve things a bit, that's not the main goal. This is
a pre-requisite for implementing keyset iteration support in JIT, as these
implementations must match.

Reviewed By: mdko

Differential Revision: D57419094

fbshipit-source-id: a97d60256a7bc569bfaede43d17912c48be33bf3