Provide dynamic dst type for LdPtrIterKey and LdPtrIterVal
Summary:
These two instructions are used at the loop header, which has to clear frame
state for locals using clearForUnprocessedPred(), as there might be a different
version of them coming via IterNext.

That means we lose track of the array specialization except of the part that we
guarded for. Constant arrays are quite common, but we don't guard them, so at
the time LdPtrIterKey/LdPtrIterVal is gen'd, we don't know the type. Same for
repo auth types coming from hhbbc.

Once irgen is done, we reflow these types and that's when more precise array
type information gets known to the loop body.

Make these instructions define destination type based on the iterated array, so
that the type gets improved during reflowTypes().

Also fix a bug in keysetPosType(). Without this, it doesn't do much for keysets.

Reviewed By: mdko

Differential Revision: D57426369

fbshipit-source-id: d765c06ee42ee4edd9b0fa06cf175d7965f4ae26