Save memory in intern builds the better way
Summary:
Reduce the amount of memory needed for trace scheduling with intern
builds. Instead of disabling trace scheduling above an arbitrary limit
(which was a quick fix), rework some of the data structures to be
smaller. Use vectors instead of maps in a few places. Instead of using
multiple maps, have a single map mapping a name to a set of
properties. Also add some missing moves. Use a latch to avoid the
co-routines queueing up while holding a lot of memory.

This all can be improved even further by re-designing the algorithm
and data-structures to take advange of common dependency sequences,
but this is fine for now.

Reviewed By: mdko

Differential Revision: D57455724

fbshipit-source-id: a7fa3b897a73ceddc8f170c9efe3cf55ab0c8b1a