Switch logging to ServerInit instead of ProcessInit
Summary:
Logging to the hhvm_runtime_options scuba table relies on `StructuredLog::*`,
which in turn relies on the scribe extension being initialized
(`StructuredLog::enabled()` needs to return `true).

The ext_scribe in turn initializes using `ProcessInitConcurrent`, which races
with `ProcessInit`. This means there's not guarantee that all we log
hhvm_runtime_options for all machines that we intend to.

Move the logging hook to ServerInit, which runs after ProcessInitConcurrent
completes to fix the race condition.

Reviewed By: ricklavoie, binliu19

Differential Revision: D57308778

fbshipit-source-id: e7ab8525e5f749f978f1058627b043066d2d48fe