Do not set the request context in ViaCoroutinePromiseBase::scheduleContinuation()
Summary:
Setting the `RequestContext` is expensive, especially if the context has registered hooks. If we do it both before submitting work to the executor, and in the scheduled callback, we're essentially paying double the cost for the cases where the request context is not already set.

Also, we're not doing it consistently across our async frameworks (`Future` doesn't do this), and adding it to `Future` would likely be prohibitively expensive, so it is better to remove this for now and explore a different solution if we need the executor to be always aware of the request context.

Differential Revision: D57275308

fbshipit-source-id: 5aa2bec88e468223dc12d615dd9e28d8c04f6faa