CursorSerializer 9/n: Support writing strings of unknown size
Summary:
This is a niche API to allow zero-copy use of libraries that perform string manipulations including compression and charset transcoding that change the size of a string in a bounded but unpredictable way.
It provides direct access to a contiguous buffer that will hold the final string value and machinery for updating the size afterwards. The caller is responsible for determining an appropriate max size as well as the correct final size.

Reviewed By: vitaut

Differential Revision: D57131341

fbshipit-source-id: afbdc8e756777f56c86071ee099ff7044f700b54