CursorSerializer 8/n: Introduce StructuredCursorWriter
Summary:
Entry point for cursor serialization.
There's a fair bit of complexity in this diff related to filling in default values for skipped unqualified fields. We are matching the behavior of the generated code, which means we write all default values for unqualified fields and custom default values for terse fields if the user does not provide a value. However, we also have to write them in field id order. Our reflection library only allows accessing fields in declaration (ordinal) order, so we add some machinery to bridge this gap.

Reviewed By: vitaut

Differential Revision: D57130873

fbshipit-source-id: 727866d5fcd26f36191931b73313a05bb0ad5256