Add raw attribute value to attributes
Summary:
Attributes contain aribtrary static expressions. Those expressions can be nested and include nodes that we do not want to parse in the fast decl parser.
We can still return the raw string value of the attribute so we could either use it in hack with existing custom parsing or pass it back to higher fidelity parsing via an additional api call.

Reviewed By: patzar

Differential Revision: D56884916

fbshipit-source-id: ecb777782e6517c47ca4f3836c87a3cb71530f62