Refine and filter bespoke sink layouts by the known layout
Summary:
We might profile that a given array sink is likely to see a particular array
layout, but we might also know something about the actual layout we are
getting and that might contradict or refine the prediction.

So, take advantage of that knowledge and avoid emitting dead checks that always
side exit.

Differential Revision: D57518393

fbshipit-source-id: ddfb68597472d8575d1e29b4fcdfbdd94421a153