Run full DCE between GVN and refineTmps
Summary:
Yes, it is valuable to run refineTmps after GVN. But it is catastrophic to do
so without a full DCE. GVN leaves a lot of dead Movs around and refineTmps
happily rewrites all references to the original SSATmp with the dst of these
Movs. This tends to create a lot of phis that do not get cleaned up if they are
part of cycles.

Reviewed By: mdko

Differential Revision: D57529413

fbshipit-source-id: 2d520dda157993e7031b98ec3904423958c10aa3