Fix findDominators() algorithm
Summary:
This cannot be a dataflow algorithm -- just because an immediate dominator of a
node did not change does not mean anything. Our successor might still need to
be recalculated and use our non-immediate dominators that might have changed.

Let's just keep rerunning the whole algorithm until convergence, which is what
the standard algorithm for this problem does.

If this becomes too slow, we might consider implementing Lengauer-Tarjan with
path compression, or one of the more recent complex modifications with
guaranteed linear complexity.

Reviewed By: mdko

Differential Revision: D57539398

fbshipit-source-id: acd7f05b90fd310abfdb89588e7eeeb7dc29e24d