refactor sym_def.t
Summary:
`Sym_def.t` is defined as
```
type t = {
  kind: kind;
  name: string;
  full_name: string;
}
```
where `full_name` is equal to `name` for top-level types, and is of the form `container::member` for members.  Which is error prone and force clients to parse the full name. This diffs turn this into a algebraic type where the right structure is exposed for each of the kinds.

Follow-up: `Index_xrefs` can be simplified furthermore

Reviewed By: pepeiborra

Differential Revision: D57538737

fbshipit-source-id: 069a43187114fc108a5fd23be42e16f17fb307f9