Deshim atomic_read_mostly_main_ptr in folly
Summary:
The following rules were deshimmed:
```
//folly/experimental:atomic_read_mostly_main_ptr -> //folly/concurrency/memory:atomic_read_mostly_main_ptr
```

The following headers were deshimmed:
```
folly/experimental/AtomicReadMostlyMainPtr.h -> folly/concurrency/memory/AtomicReadMostlyMainPtr.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

Reviewed By: Orvid

Differential Revision: D57681720

fbshipit-source-id: b460bf617837a51d59a1b795c6b2da6db5871874