Auto-chunk HTTP/1.1 responses even without keepalive.
Summary:
Even though we can't reuse the connection, chunking still
prevents us from depending on the TCP/TLS connection closure to indicate the
end of response. This is benefical, because it prevents truncation attacks when
the response is sent over TLS, but the client doesn't enforce receipt of the
TLS close_notify.

Reviewed By: afrind

Differential Revision: D57452492

fbshipit-source-id: 9119514a37bd8a64da07bdecbba62bcf995672b0