Relocate mstch cache into lambda capture, instead of separate unordered_map
Summary:
This is a refactor I need for efficiently implementing D57690617.

If you are wondering why a cache is needed here despite being unconditionally overwritten on every call (which D54435037 describes as *"broken caching behavior in the thrift mstch-based code-generators"*) it is because we need `object_t::at` to return `const N&`, not `N`, in order for classes like `get_token` to be usable as a `std::variant` visitor without doing large copies in the non-object branches. In Rust we might use `Cow` in this scenario, but that gets complex in C++ (`std::variant<N, const N&>`).

https://www.internalfb.com/code/fbsource/[f782a6966964cd2a97aebf96e9e84569bceb71bb]/fbcode/thrift/compiler/detail/mustache/visitor/get_token.h?lines=43-52

I verified that this diff is performance-neutral using a few large Thrift targets in `rust` and `cpp2` and the methodology from D57690617's test plan.

Incidentally, I found it is not performance-neutral if `register_methods` delegates to `register_method`, as that doubles the number of `std::function` allocations compared to the original code.

Reviewed By: iahs

Differential Revision: D57694787

fbshipit-source-id: 5e580d33a8273a78e586d998443aaa9e661551e5