Deshim flat_combining in folly
Summary:
The following rules were deshimmed:
```
//folly/experimental/flat_combining:flat_combining -> //folly/synchronization:flat_combining
```

The following headers were deshimmed:
```
folly/experimental/flat_combining/FlatCombining.h -> folly/synchronization/FlatCombining.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

Reviewed By: Orvid

Differential Revision: D57686159

fbshipit-source-id: fe76a1f716875b844799a15d4cc58a0410444716