Weaken the this type in refinement
Summary: Currently the use of `as this` and `is this` is unsound, unless the enclosing class is final and non-generic. We could ban the construct, but there are no good local workarounds. So instead, as with intish cast and enums, we can treat the type of `as this` as `~this` rather than the unsound `this`.

Reviewed By: vassilmladenov

Differential Revision: D57677289

fbshipit-source-id: 1c9f5d9000ffcfcc59856cef4951bf1abaffe5e0