Always read strings into string_view when Contigious = true
Summary:
We require the user to tell us whether a given cursor looks into a contiguous buffer to enable the read overload that modifies a string_view. Given that we have that information available there isn't a good reason to read into std::string when Contiguous is true; the user can perform the conversion as cheaply as we can construct the string, and often a view is sufficient.
This is most salient for the container iterator API, where there is currently no way to read into a string_view (i.e. without allocating a string). But once we're doing it there we may as well do it everywhere.

Reviewed By: yoney

Differential Revision: D57595293

fbshipit-source-id: a95a28005c690febfa979199a65bd9c4f1660c10