Support cursor read inside containers
Summary:
Allows efficient reading of containers and structured types inside containers.
The iterator API is kept as the only way to read containers of primitives, and continues to support all types when allocating elements is acceptable.
To support this flexibility the initial element read is moved from the constructor to the `begin()` call.

Reviewed By: yoney

Differential Revision: D57697927

fbshipit-source-id: 1e4da2468041afb83c21223c1e9d24a62fd16908