Deduplicate loads of keys and values in keysets
Summary:
When iterating keysets with both keys and values (such as in Dict\map), we load
both of them using LdPtrIterVal. This instruction does not currently support
GVN, as arrays in SSATmps might be still mutated without creating a new SSATmp.

Solve this problem at the root -- when iterating by key-value pair, emit only
one LdPtrIterVal instruction and use it for both key and value.

Differential Revision: D57793176

fbshipit-source-id: 27b1ad388a6b586a24a070dcf83e3e1b9311d803