FOLLY_ATTR_MAYBE_UNUSED_IF_NDEBUG
Summary:
Macro `FOLLY_ATTR_MAYBE_UNUSED_IF_NDEBUG` expands to `maybe_unused` when `defined(NDEBUG)`.

Example:
```
[[FOLLY_ATTR_MAYBE_UNUSED_IF_NDEBUG]] int a = 0;
// ...
a += 1;
// ...
assert(a==3);
```

This way, when `defined(NDEBUG)`, the variable `a` is marked `[[maybe_unused]]` and so the lack of use will raise no warning. But when `!defined(NDEBUG)`, the variable is not marked, and accidental lack of use will raise a warning.

Suggested by r-barnes.

Reviewed By: r-barnes

Differential Revision: D57801880

fbshipit-source-id: 8830e9ad2a6d8762dcc3d48531b02c99658e8230