Add JIT support for iteration over tombstones
Summary:
Instead of bailing out to the C++ helper if we profiled there are tombstones,
emit code to handle tombstones if they could be present.

- vecs: unaffected
- dicts: translate code even if profiled key types contain tombstones, then handle them
- keysets: we don't profile keys; don't guard for tombstones, just handle them
- bespoke: continue guarding for tombstones if they could be possible due to configuration, which they aren't possible with default config

There are two separate emitted loops for skipping tombstones, one in IterInit,
another in IterNext. Each of them adds just a couple of asm instructions. It
would be possible to instead perform the check in the loop header and jump to
the IterNext on tombstone, but that would add a significant amount of
complexity (e.g. IterInit would need to have third destination).

After this diff, JIT should handle almost all array iteration, except when
profiling says we have a mix of vanilla and bespoke arrays, which is rare.

Reviewed By: mdko

Differential Revision: D57809680

fbshipit-source-id: 87488e02a4e52680e61c9f9b0927be905e11e744