Revert D55388400: Multisect successfully blamed "D55388400: [mysql/8.0.32] Replace 8.0.20 fbmysqlclient with mysql/8.0.32 fbmysqlclient" for one test failure
Summary:
This diff reverts D55388400
D55388400: [mysql/8.0.32] Replace 8.0.20 fbmysqlclient with mysql/8.0.32 fbmysqlclient by hermanlee causes the following test failure:

Tests affected:
- [cogwheel:cogwheel_model_master_asan_test_changes#test_model_master_sm](https://www.internalfb.com/intern/test/562950077981591/)

Here's the Multisect link:
https://www.internalfb.com/multisect/5194449
Here are the tasks that are relevant to this breakage:

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: hermanlee

Differential Revision: D57813150

fbshipit-source-id: fd113e0350adc6717577e81874d49a795f66a3de