bit stuff... Fix conversion of error hash sets between rust and ocaml
Summary:
Rust to Ocaml int conversion will do:
```
x << 1 | 1
```
then Ocaml will disregard the LSB and treat the rest as a 63 bit signed integer. That means that the sign bit in the rust isize will be discarded and the second most significant bit in the rust isize will become the sign bit in OCaml. So this conversion is only safe to do (i.e. will not change which integer is represented) if and only if the two most significant bits of the rust isize are the same.

If that's not the case, spooky things can happen. In our case, we had a `BTreeSet<isize>`, which upon conversion changed some integers but not others, meaning the BTreeSet was not ordered anymore, meaning when converting that set to OCaml, the ISet was also out of order.

To fix that, we take the result of error hashing and flip the MSB so that it's equal to the second MSB.

In a future diff, we'll look at making ocamlrep safer in general, and possibly fix other occurrences, like with DepSet.t.

Differential Revision: D57723495

fbshipit-source-id: 94b60da3eb190fcb26f48f9633ee059cc55c9ce5