change Classname to HH::classname in Textual
Summary:
@public
Classnames in parameters were given type hint `Classname` whilst type structures used in dynamic type tests used the name `HH\\classname`, so newly added runtime type tests were failing. This unifies them and adds `extends HH::classname` to the `C$static` types we use for the companion objects that represent classnames in InferHack, provided they don't already extend something.
Note that this isn't the nicest solution: it would be better if hackc translated just this instance of a hack generic type into different Textual types, viz `classname<A>` goes to `A$static`, but I'm not sure if this is all going to need another revision to make the translation of classnames compatible with `string` in the near future, so haven't done that yet

Reviewed By: geralt-encore

Differential Revision: D57841985

fbshipit-source-id: 8501b906d1cef15e3fca1b9545cf85b3df5294dd