Disable unreachable code warning in folly::expected::thenOrThrow_() in MSVC
Summary:
In a similar vein to D55028245, the thenOrThrow_() functions in folly::expected are causing unreachable code warnings in MSVC which block compilation in builds that treat warnings as errors.
````
xplat\folly\Expected.h(711) : error C2220: the following warning is treated as an error
xplat\folly\Expected.h(711) : warning C4702: unreachable code
xplat\folly\Expected.h(711) : warning C4702: unreachable code
xplat\folly\Expected.h(711) : warning C4702: unreachable code
````

This happens only in MSVC when a caller uses a folly::expected::as{Type}() function where the compiler knows through optimization that the expected will always hold a valid value (thus the exception-throwing code can never be reached).  Issuing a warning in this case seems to be an overreach by the MSVC compiler but apparently is by-design.

This is forcing HSR to add warning disable pragmas anywhere that folly::dynamic is used as it triggers these warnings.  Better to fix the warning at the source so it doesn't keep biting us.

Reviewed By: Orvid

Differential Revision: D57844817

fbshipit-source-id: 60dba228788dbd1b47d71382b4487b7cc1440dd1