Fix package v2 tests
Summary: The previous regex matching implementation would return True trivially so it wasn't performing the package v2 tests correctly. In D57886539 I changed it such that with the flag --normalize-paths, the verifier would only strip the cwd from filepaths referenced in the actual error message and compare it against the expected as normal.

Reviewed By: DavidSnider

Differential Revision: D57886540

fbshipit-source-id: d2dce10e42785f5829fcc52cde558bd4b836caf9