Fix/change --check-regex-match in verify.py
Summary:
I added this option in D53027218 for a package v1 feature which could emit error messages containing filepaths, specifically, errors of the form
```
error: Typing[4472] Cannot access a public element which belongs to package pkg2 from package pkg1 [1]
-> This is from FILE2, which belongs to package pkg2 [2]
-> But FILE1 is in package pkg1, and package pkg1 does not include package pkg2 [3]
```
the paths of FILE1 and FILE2 can vary depending on the environment in which the typechecker is run, hence the need for regex match.

But as it turns out, the regex match wasn't correctly implemented and really, harvesting the paths after the current working directory should suffice. Additionally, since the package v1 feature has been killed, and this option is only used to now verify package v2, rename it to reflect so.

Reviewed By: CatherineGasnier

Differential Revision: D57886539

fbshipit-source-id: 35eedbcaaa5c50a33a6ce78c35563ebb016f88fb