warning hashing ignores error lines
Summary:
High-level goal: we want to be able to hide pre-existing warnings. We get the hashes of pre-existing warnings from the saved states. However, if a pre-existing warning moves around because the user adds lines, its hash will change and therefore will not be filtered out anymore.

To remediate that, we change `UserError::hash_for_saved_state` so that it first sets all line numbers in the error to 0. Actually, that's not quite true: we change all position spans to pretend they start at line 1, but keep the span length unchanged.

We've discussed converting positions to be relative to the method signature. However, that means changing a method name will make its pre-existing warnings appear, so that's not really satisfactory. It's also much more complicated to implement.

I think that line-agnostic hashing is good enough because other distinguishing features will be hashed: exact error message which usually contains type or def names, start and end columns.

Reviewed By: madgen

Differential Revision: D57910940

fbshipit-source-id: 9fdde5381bada4813c23373ae83a76f58b28f8b4