Fix supertype positions under extended-reasons
Summary:
The relevant position for the reason of the super-type is the rightmost reason of the flow but the `to_raw_pos` function will find the leftmost reason. This leads to the incorrect position being reported for the 'expected' type.

This diff reverses the flow for the supertype before reporting its position so the rightmost reason becomes the leftmost reason.

Differential Revision: D57621202

fbshipit-source-id: 8c230cf44a0b0137f870d5f1448c307b68040ce4