Remove redundant flow for function args
Summary:
Removes unnecessary flow from expressions used in function argument position into a witness for that function argument which leads to the type of a sub-expression seemingly flowing into itself.

Orginally, the intention  was to use more explicit witnesses to make human readable paths clearer but this is already accounted for by the associated projection.

Reviewed By: madgen

Differential Revision: D57552933

fbshipit-source-id: b3c4a0b0ec5d822cbe1332f124249298f6a63965