Retain reasons through tyvar expansion
Summary:
Currently when we expand a type variable, we also discard the reason, using both the type and the reason of the type we expanded to.

Under extended-reasons we change this behaviour so that we don't loose provenance; to do this, we record a flow from a type variable's solution into the type variable when we expand it and that expansion results in a concrete type.

NB, the typing of `Destructure` is a bit off at the moment, so we end up duplicating provenance for this case. I'll address this in a later diff

Reviewed By: madgen

Differential Revision: D57207264

fbshipit-source-id: 390829f50f4249a4c6852c03c4cc11582e081de5