Add new config option to put buggy paths under
Summary:
The typechecker relies on reasons for both error reporting and, in some cases, typing. Further, the error reporting implementation interacts with FIXME error silencing in a brittle way. Because of this, some of the places we want to introduce extended-reasons lead to peturbations of errors or even spurious errors.

However, in practice these spurious errors are fairly limited in scope. This diff adds a new config option (`yolo`) which lets us record the provenance on buggy paths.

Differential Revision: D58009830

fbshipit-source-id: ce07ca34b4a126762432c282a4a79267680542a4