Adjusting the WritableIOBuf to be able to take an int for initialization
Summary:
We want to be able to use C++ to initialize the buffer, to increase the speed of allocation. This seems to be primarily from how Python appears to allocate and zero out the bits in the buffer.

This diff passes along the capacity to the C++ code to initialize the buffer.

Additionally, I changed some of the ways for how we initialize the WritableIOBuf as length & capacity are not the same thing. From the documentation, this is a diagram for the difference.
```
 *      +-------+
 *      | IOBuf |
 *      +-------+
 *       /
 *      |            |----- length() -----|
 *      v
 *      +------------+--------------------+-----------+
 *      | headroom   |        data        |  tailroom |
 *      +------------+--------------------+-----------+
 *      ^            ^                    ^           ^
 *      buffer()   data()               tail()      bufferEnd()
 *
 *      |----------------- capacity() ----------------|
```

If you treat them as the same, when grabbing the bytes for an empty buffer, you get some gibberish for unitialized bytes `\xbe`. However, as my diff makes the change to use them according to the IOBuf spec, that no longer happens.

However, due to that change, we now need to account for management of the buffer pointers. So, this added a few of the ways that we can adjust the pointers (`append`, `prepend`, `trimStart`, `trimEnd`).

Ultimately, this seemed to be the correct tradeoff, in order to keep getting what are expected results. The next diff then shows this in action, and how we can use this to optimize Manifold's read operation.

Reviewed By: mofa28

Differential Revision: D57890817

fbshipit-source-id: 009b6e53700c3456a9dd6b69d554320cb5ff409a