disallow invoking protected methods on interfaces
Summary:
Invoking a protected method on an interface can cause a fatal at runtime since the receiver might be a class that implements the interface that is unrelated from the class hierarchy of the caller (see test `protected_in_interface2.php` below).

This diff forbids invoking a protected method on a receiver that has an interface type.

Reviewed By: madgen

Differential Revision: D57045251

fbshipit-source-id: c1503ebef30b445b12a07c60bc152c7ff18746b9