Adding a validation endpoint to facts, currently just checking for duplicate facts definitions in sqlite db
Summary:
Facts database can currently lose integrity and may be invalid. In such a case we found that the same type was pointing to 2 different paths.

This diff adds a validation endpoint -> HH\Facts\validate(); that will currently throw an exception if any types lead to duplicate paths.
This can be further extended in the future for more queries that we know invalidate facts.

This endpoint also takes in an optional vector<string> of types to ignore if we want to override to not fail on certain classes we KNOW is currently a problem but not able to fix yet

Reviewed By: paulbiss

Differential Revision: D57303153

fbshipit-source-id: 72a11b9125485a10804a236475c81a0dfce3eb10