make_exception_ptr_with
Summary:
Add `folly::make_exception_ptr_with` like `std::make_exception_ptr`, constructing an instance `std::exception_ptr` efficiently.

Overloads:
* A single invocable argument. The argument is invoked and its return value is the managed exception.
* Two arguments, the first of which is `in_place`. The argument is moved or copied copied and the result is the managed exception.
* Variadic arguments, the first of which is `in_place_type<E>`. An exception of type `E` is created in-place with the remaining arguments forwarded to its constructor, and it is the managed exception.

This may be more efficient than throw-and-catch-all with `std::current_exception()`, which involves going into the unwinder (`libunwind`), as well as the dynamic loader (`libdl`) to hold a global lock during unwinding. This depends on the C++ implementation having a way to avoid throw-and-catch-all, even if that way is not public.

This may be more efficient than optimized `std::make_exception_ptr`, which libstdc++ has had for a decade and which libc++ just got this year, in cases where the exception object itself is costly to move. This is the case for most standard-library exception types, for which move is just copy and for which copy is an atomic-rmw refcount increment and a corresponding atomic-rmw refcount decrement. This depends on the C++ implementation having a way to avoid copy, even if that way is not public.

Reviewed By: Gownta

Differential Revision: D57677424

fbshipit-source-id: b370fbc766aed1887c01c384f9a74d11a7810fed