Always record error when using `|||` operator
Summary:
When using the `|||` operator with two propositions that are invalid, we will call `disj` with `fail:None` and end up constructing `Disj(None,[])`. When this gets to `Subtype_trans.tell`, we drop the error on the floor.

This diff reintroduces a `fail` argument to that operator to ensure we record an error in this case.

Differential Revision: D58137268

fbshipit-source-id: 8e019c0fda5f8822df36c6edc9df60c252825349