Include errors for previous disjuncts which were found to be invalid
Summary:
The `Disj` data constructor for `subtype_prop` encodes both invalid propositions _and_ disjunctive propositions whose validity cannot yet be determined.

In the second case, we may well have determined that a number of the other branches were invalid and then put the remaining branches back into a disjunction until we can determine their validity via a transitivity step.

This diff updates the handling of disjunctions in the transitivity step so that we combine the errors on definitely invalid branches with the errors on branches found to be invalid only after a transitivity step and report them all if we have no valid disjunct.

Differential Revision: D58182636

fbshipit-source-id: b5a1edf35f350d4d691d8e58fc7529b7e12df0a1