Move PrimaryPtr to fbcode/folly/concurrency/memory
Summary:
The following targets were moved to folly/concurrency/memory:
```
//folly/experimental:primary_ptr
//folly/experimental/test:primary_ptr_test
```

`arc f` was applied

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:ad_mepp

Reviewed By: Orvid

Differential Revision: D57929837

fbshipit-source-id: 2b24ef5db07d00c43552ef41aec8df42a62d4cdb