Move Cleanup to folly/futures
Summary:
The following targets were moved to folly/futures:
```
//folly/experimental:cleanup
//folly/experimental/test:cleanup_test
```

`arc f` was applied

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:ad_meclen.folly

Reviewed By: ot

Differential Revision: D58024166

fbshipit-source-id: 65701585d243ec55be1799f8b1e2c490a6bfab65