Deshim FunctionScheduler in squangle
Summary:
The following rules were deshimmed:
```
//folly/experimental:function_scheduler -> //folly/executors:function_scheduler
```

The following headers were deshimmed:
```
folly/experimental/FunctionScheduler.h -> folly/executors/FunctionScheduler.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:ad_defsx.squangle

Reviewed By: jkedgar

Differential Revision: D58155130

fbshipit-source-id: d23dcf1fae8d6010bea07bba0aac3d66253173e5