Move settings/ to fbcode/folly/settings
Summary:
The following targets were moved to folly/settings:
```
//folly/experimental/settings:settings
//folly/experimental/settings:types
//folly/experimental/settings:immutables
//folly/experimental/settings/test:settings_bench
//folly/experimental/settings/test:settings_test
```

`arc f` was applied

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:ad_mes.folly

Differential Revision: D57937846

fbshipit-source-id: 7669ebec892c0ba275d6e60eca72ab842f8554e8