Synchronized ThreadEntrySet and reducing the uses of global meta lock while operating on allThreadEntryMap_
Summary:
`folly::Synchronized<std::unordered_map<uint32_t, SynchronizedThreadEntrySet>> allThreadEntryMap_;`

Avoids the use the global locks when operating on  the allThreadEntryMap_ and ThreadEntrySet and instead rely on folly::Synchronized rlock() and wlock().

Consequently reduces use and contention on the global meta lock to a large extent especially when there is a lot of thread local churn.

Reviewed By: yfeldblum

Differential Revision: D57720792

fbshipit-source-id: 8122a982bae17703d557cfdd8a31febd28fe21dc