Remove unneeded string conversions of primitives
Summary:
Rationale: just tidier code at all these call sites. I was not able to measure any performance difference from this change. The same integer-to-string conversion happens either way, it's just a matter of whether it happens with a `std::to_string` or `fmt::format("{}", …)` at the time of construction of the `mstch::node`, vs a `fmt::format_to` during rendering.

https://www.internalfb.com/code/fbsource/[0e2272aad4fb00d28440c77135230f703c7fbf77][outline]/fbcode/thrift/compiler/detail/mustache/visitor/render_node.h?lines=42%2C49-51%2C82

Either way no memory allocation is happening because the longest value of `int` is only 11 characters which is less than the inline capacity of `std::string`.

The most likely reason that a performance regression would occur is if there is some node which is rendered disproportionately many times. Before, the int-to-string formatting would have been done exactly once, whereas now it'll happen as many times as the node is rendered. To the extent that this happens in any of our templates, formatting an integer to string is cheap enough that the difference is apparently negligible in practice.

This change does affect the behavior of `is_node_empty`, which is invoked when a template does `{{#method}}...{{/method}}`. A node created using `std::to_string(0)` would be considered truthy whereas directly using integer `0` is falsey.

https://www.internalfb.com/code/fbsource/[0e2272aad4fb00d28440c77135230f703c7fbf77][outline]/fbcode/thrift/compiler/detail/mustache/visitor/is_node_empty.h?lines=37%2C46%2C52%2C55

However, this doesn't impact the behavior of any existing template because there would have been no reason for any of them to be doing a `{{#...}}` on a node that holds a stringified integer, as such a node would never have been falsey before this diff.

Reviewed By: diliop

Differential Revision: D57946615

fbshipit-source-id: 5a3d04bc63418018347afb344b6b649b83bf9072