Delete {{value:value}} method
Summary:
This method is used only in a single template, and has surprising semantics, and confusing usage.

It caught my attention because of this code:

https://www.internalfb.com/code/fbsource/[f211a5488581e8d7baa112cb81f85259bb23893e]/fbcode/thrift/compiler/generate/templates/json/const_value.mustache?lines=19-22%2C43

In pseudocode the above is equivalent to:

```
if (!value.is_string()) {
    if (!value.value().is_empty()) {
        if (value.is_string()) {
            render('"');
        }
        render(value.value());
        if (value.is_string()) {
            render('"');
        }
    }
}
```

The meaning of `value` does not change throughout this code, and no mutation is happening within any of the methods. If we know `value.is_string()` is **false**, why are there cases appearing to handle `value.is_string()` being **true**?

I came up with a guess of how this code ended up like this, and it has to do with `{{value:value}}` being a bad API.

At a high level, the template in const_value.mustache is trying to handle 3 cases:

- If the value is a primitive double/bool/integer/string/enum
- If the value is a list
- If the value is a map

https://www.internalfb.com/code/fbsource/[f211a5488581e8d7baa112cb81f85259bb23893e]/fbcode/thrift/compiler/generate/templates/json/const_value.mustache?lines=20%2C22-23%2C29-30%2C42

But `{{#value:value}}` does not mean `if (!value.is_list() && !value.is_map())`. It instead means `if (!value.value().is_empty())`. And for mstch::node containing a string, `value.value()` is the value of the string. A node containing an empty string would hit none of the 3 cases above!

So the author had to add a whole separate case in order to get empty strings to render as intended (`""`).

https://www.internalfb.com/code/fbsource/[f211a5488581e8d7baa112cb81f85259bb23893e]/fbcode/thrift/compiler/generate/templates/json/const_value.mustache?lines=18-20%2C22-23%2C29-30%2C42-43

This is an example of `{{value:value}}` working against us. It's used in one place, and is not especially well suited even for that place.

And why is it not used anywhere else? Because none of them want `{{value:value}}`'s behavior on booleans, which it inherits from `std::to_string`. Booleans get turned into `0` and `1` by `std::to_string`, whereas every other Thrift output language deals with booleans as `false`/`true` or `False`/`True`.

Also because handling the non-list non-map cases the straightforward way is straightforward.

https://www.internalfb.com/code/fbsource/[f211a5488581e8d7baa112cb81f85259bb23893e]/fbcode/thrift/compiler/generate/templates/rust/lib/value.mustache?lines=29-32%2C35

This diff deletes `{{value:value}}` and replaces its 1 use with a more direct transcription of the behavior that was intended.

Reviewed By: diliop

Differential Revision: D57946612

fbshipit-source-id: 600183837f3d8b5b0bb107909c2f908deea16205