Support sorting keys in json_encode
Summary: I have some Hack code which serializes and hashes Hack arrays, and then uses the hash as a cache key. It also needs to deserialize this string back into its original representation (or an equivalent). It currently uses JSON, but it seems there isn't a serialization method that deterministically orders array keys (i.e. `dict['a' => 1, 'b' => 2]` and `dict['b' => 2, 'a' => 1']` serialize to different strings). This diff adds an option to do that in VariableSerializer, and then a flag to json_encode. Also posted about this here: https://fb.workplace.com/groups/hackforhiphop/posts/26327299490225317.

Reviewed By: ricklavoie

Differential Revision: D58189573

fbshipit-source-id: 43a2425b5590e562b9bc1e8574ee08da911a77af