Make JIT agreee with VM for `is_a`
Summary:
The JIT implementation of `is_a` and `is_subclass_of` does not honor the intrinsic's third argument (`$allow_string`), so it returns `true` in cases where the VM implementation would return `false`. It also didn't check for traits and would not jit in the case of string or lazy class for the first argument. This discrepancy didn't cause an issue because before `HH\classname_to_class`, there was no way to get a TCls value

- Simplified and made vm and jit impls match
- Added silent fallback to LdCls[Cached] and made them return S(Cls|Nullptr)
  - Similar to D40613184, extended `AssertNonNull` to support Cls
- Note: we do a `Class::load` on the first argument and `Class::lookup` on the second
- Unblocks {D58048561}

Future work:

The __Native func forces the second argument to be a string, leading to unnecessary implicit conversions. Will expand it to the `classname` union and `spatch` the WWW callsites to get rid of the strings.

Reviewed By: voorka

Differential Revision: D58073164

fbshipit-source-id: 06dda3c80dc69a384e00ba3fb02228775e10f7d9