Don't instantiate EventHandlers for invalid events.
Summary:
Rather than defining a catch all EventHandler::handle(...) defined for each invalid (state, event) pair that delegates to the `InvalidEventHandler`, lift up the "invalid event handler" dispatching logic to `getEventHandler` and leave `EventHandler` undefined for invalid events and use SFINAE to dispatch to the correct one.

This has the effect of not instantiating a bunch of EventHandler::handle symbols that all simply call InvalidEventHandler.

From a .text size perspective, this saves 1584 bytes of .text.

But where we see this effect the most is in the final outputted .so size, which is reduced by 7920 bytes. This is mainly due to the fact that we instantiate fewer distinct functions, which reduces the size of our DWARF unwinding tables in .eh_frame.

Reviewed By: knekritz

Differential Revision: D58151754

fbshipit-source-id: 8614048a093f99a56ef5c3616bfdaf8de047e932