Emit lint warning for redundant generics with multiple as bounds if the bounds simplify to a single type
Summary:
The introduction of implicit pessimisation led to us missing lint warnings on redundant generics with `as` bounds, for example:
```
function foo<T as arraykey>(T $x):void { }
```
Under pessimisation this becomes
```
<<__SupportDynamicType>>
function foo<T as arraykey as supportdyn<mixed>>(T $x):void { }
```
The pessimisation process doesn't simplify the bounds, and the redundant generics linter won't warn that `T` is redundant if it has multiple `as` bounds because they must be expressed as an intersection. But when the intersection simplifies to a single type, as here (just `arraykey`), we should still emit the warning i.e.
```
The generic parameter `T` is redundant because it only appears in a contravariant (input) position. Consider replacing uses of generic parameter with `arraykey` or specifying `<<__Explicit>>` on the generic parameter (Lint[5624])
```

Reviewed By: madgen

Differential Revision: D58276808

fbshipit-source-id: 4d162a706d2698f714bf65ba03086da1303cfe9b