Support await in splices
Summary:
Allow `await` in splices in expression trees. This was mostly a matter of turning off the error, since the expression tree gets desugared to code that the await lifting code can understand.

The immediately called lambdas that get introduced by the expression tree desugaring need to be marked as async; otherwise, it cannot contain an await. We add tracking in the lowerer to determine if an expression tree contains an await-in-splice, and is inside of an async function. Only then do we introduce an async lambda in the desugaring. Otherwise, we stick with a synchronous one.

Marked as an Unstable feature as it isn't thoroughly tested yet.

The lowerer had a `map_yielding` function that managed tracking whether a function body contained a `yield` expression, using some state in the Env. This renames it (and moves it to the Env's impl) and generalises to tracking `await` expressions (also bottom up), and whether the expression is inside an `async` function, which is top down.

Reviewed By: dlreeves

Differential Revision: D58018447

fbshipit-source-id: b3a8c253a028976ee4aa9c26562899819aa1e321