Detect non-int access to `vec`s
Summary:
This strenghtens the invalid index check by accounting for `vec`s.

This should be done in core type inference, but it cannot be because `vec` keys are covariant, so the typechecker is respecting the spec when it lets string indexing on `vec`s.

In fact, all the key indexing errors for a `vec` indexed by a `string` are (sometimes deliberate) incompleteness.

Reviewed By: andrewjkennedy

Differential Revision: D58052037

fbshipit-source-id: d8c42ea0232d226185d04664dd292417f2ff37db